Remove top-level re-export files
