+++ tests/agents/test_generator_tools.py
@@ -8,8 +8,8 @@ from typing import Any
 import pydantic
 
 import ai
-from ai import events as agent_events_
 from ai import models
+from ai.types import events as agent_events_
 from ai.types import events as events_
 from ai.types import messages as messages_
