+++ tests/agents/test_hooks.py
@@ -10,7 +10,7 @@ import pydantic
 import pytest
 
 import ai
-from ai import events as agent_events_
+from ai.types import events as agent_events_
 
 from ..conftest import MOCK_MODEL, mock_llm, text_msg
