+++ tests/agents/ui/ai_sdk/outbound/test_sse.py
@@ -3,9 +3,9 @@ from __future__ import annotations
 import json
 from collections.abc import AsyncGenerator
 
-from ai import events as agent_events_
 from ai.agents.ui.ai_sdk import protocol, to_sse
 from ai.agents.ui.ai_sdk.outbound.sse import format_sse, serialize_part
+from ai.types import events as agent_events_
 from ai.types import events as events_
